requirements.txt
# Core ML/DS Libraries
accelerate
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
transformers
huggingface-hub
safetensors
sentencepiece
tokenizers
numpy==2.3.0
pandas
pyarrow
Pillow==11.2.1
scikit-learn 

# NLTK e TQDM
nltk==3.8.1
tqdm==4.66.4

# Django and Web Framework
Django==5.2.3
djangorestframework==3.16.0
gunicorn==23.0.0
asgiref==3.8.1
dj-database-url==3.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3

# Async and Networking
aiohappyeyeballs
aiohttp
aiosignal
async-timeout
Brotli
certifi
charset-normalizer
h2
hpack
hyperframe
idna
requests
urllib3==2.4.0
PySocks
yarl

# Utility and Others
attrs
Bottleneck
cffi
colorama
dill
filelock
frozenlist
fsspec==2025.3.0
gmpy2
iniconfig==2.1.0
Jinja2
MarkupSafe
mpmath
multidict
multiprocess
networkx
numexpr
packaging
pluggy
propcache
protobuf==6.31.1
psutil
pycparser
Pygments==2.19.1
pytest==8.4.0
pytest-django==4.11.1
python-dateutil
pytz
PyYAML
regex
six
sympy
tqdm
typing_extensions
tzdata
xxhash
zstandard